// src/shared/types/filter-tags.ts
export enum FilterTags {
    Destroyer = 'Destroyer',
    Cruiser = 'Cruiser',
    Battleship = 'Battleship',
    Aircarrier = 'Aircarrier',
    Submarine = 'Submarine',
    Commonwealth = 'Commonwealth',
    Europe = 'Europe',
    France = 'France',
    Germany = 'Germany',
    Italy = 'Italy',
    Japan = 'Japan',
    Netherlands = 'Netherlands',
    PanAmerica = 'PanAmerica',
    PanAsia = 'PanAsia',
    Spain = 'Spain',
    UK = 'UK',
    USA = 'USA',
    USSR = 'USSR',
    I = 'I',
    II = 'II',
    III = 'III',
    IV = 'IV',
    V = 'V',
    VI = 'VI',
    VII = 'VII',
    VIII = 'VIII',
    IX = 'IX',
    X = 'X'
}